Move ldap_sync tests into package
Also, move the `action`-like tests into a separate test module.
